# docs/openapi-v3.yaml
openapi: 3.0.3
info:
  title: Product Requirements Management API
  description: |
    Comprehensive API for managing product requirements through hierarchical structure of Epics → User Stories → Requirements. 
    Features include full-text search, comment system, relationship mapping, and configurable workflows.
  version: 1.0.0
  contact:
    name: API Support
    url: http://www.swagger.io/support
    email: support@swagger.io
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.requirements.example.com
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /api/v1/epics/{id}/user-stories:
    parameters:
      - $ref: '#/components/parameters/EntityIdParam'
    get:
      tags: [Epics]
      summary: Get epic with user stories
      responses:
        '200':
          description: Epic with user stories
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Epic'

    post:
      tags: [Epics]
      summary: Create user story in epic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserStoryRequest'
      responses:
        '201':
          description: User story created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserStory'
...
  /api/v1/user-stories:
    get:
      tags: [User Stories]
      summary: List user stories
      parameters:
        - name: epic_id
          in: query
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/CreatorIdParam'
        - $ref: '#/components/parameters/AssigneeIdParam'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/UserStoryStatus'
        - $ref: '#/components/parameters/PriorityParam'
        - $ref: '#/components/parameters/OrderByParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
        - $ref: '#/components/parameters/IncludeParam'
      responses:
        '200':
          description: List of user stories
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserStoryListResponse'

    post:
      tags: [User Stories]
      summary: Create user story
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserStoryRequest'
      responses:
        '201':
          description: User story created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserStory'
  /api/v1/user-stories/{id}:
    parameters:
      - $ref: '#/components/parameters/EntityIdParam'
    get:
      tags: [User Stories]
      summary: Get user story by ID
      responses:
        '200':
          description: User story details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserStory'

    put:
      tags: [User Stories]
      summary: Update user story
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserStoryRequest'
      responses:
        '200':
          description: User story updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserStory'

    delete:
      tags: [User Stories]
      summary: Delete user story
      responses:
        '204':
          description: User story deleted
  /api/v1/user-stories/{id}/acceptance-criteria:
    parameters:
      - $ref: '#/components/parameters/EntityIdParam'
    get:
      tags: [User Stories]
      summary: Get user story acceptance criteria
      responses:
        '200':
          description: List of acceptance criteria
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptanceCriteriaListResponse'
    post:
      tags: [User Stories]
      summary: Create acceptance criteria in user story
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAcceptanceCriteriaRequest'
      responses:
        '201':
          description: Acceptance criteria created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptanceCriteria'
  /api/v1/user-stories/{id}/requirements:
    parameters:
      - $ref: '#/components/parameters/EntityIdParam'
    get:
      tags: [User Stories]
      summary: Get user story requirements
      responses:
        '200':
          description: User story with requirements
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserStory'
    post:
      tags: [User Stories]
      summary: Create requirement in user story
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRequirementRequest'
      responses:
        '201':
          description: Requirement created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Requirement'
  /api/v1/user-stories/{id}/status:
    parameters:
      - $ref: '#/components/parameters/EntityIdParam'
    patch:
      tags: [User Stories]
      summary: Change user story status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatusChangeRequest'
      responses:
        '200':
          description: Status changed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserStory'
  /api/v1/user-stories/{id}/assign:
    parameters:
      - $ref: '#/components/parameters/EntityIdParam'
    patch:
      tags: [User Stories]
      summary: Assign user story to user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignmentRequest'
      responses:
        '200':
          description: User story assigned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserStory'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication. Include as Authorization Bearer token header.
  parameters:
    EntityIdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Entity ID (UUID or reference ID)
    CreatorIdParam:
      name: creator_id
      in: query
      schema:
        type: string
        format: uuid
      description: Filter by creator UUID
    AssigneeIdParam:
      name: assignee_id
      in: query
      schema:
        type: string
        format: uuid
      description: Filter by assignee UUID
    PriorityParam:
      name: priority
      in: query
      schema:
        $ref: '#/components/schemas/Priority'
      description: Filter by priority level
    OrderByParam:
      name: order_by
      in: query
      schema:
        type: string
      description: Sort order (e.g., 'created_at DESC', 'reference_id ASC')
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
      description: Maximum number of results
    OffsetParam:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
      description: Number of results to skip
    IncludeParam:
      name: include
      in: query
      schema:
        type: string
      description: Comma-separated list of related entities to include
...
  schemas:
...
    UserStoryStatus:
      type: string
      enum: [Backlog, Draft, "In Progress", Done, Cancelled]
...
    Epic:
      type: object
      description: High-level feature or initiative containing multiple user stories
      required: [id, reference_id, title, status, priority, creator_id, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        reference_id:
          type: string
          pattern: '^EP-\d+$'
        title:
          type: string
          maxLength: 500
        description:
          type: string
          maxLength: 5000
        status:
          $ref: '#/components/schemas/EpicStatus'
        priority:
          $ref: '#/components/schemas/Priority'
        creator_id:
          type: string
          format: uuid
        assignee_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        # Optional populated fields
        creator:
          $ref: '#/components/schemas/User'
        assignee:
          $ref: '#/components/schemas/User'
        user_stories:
          type: array
          items:
            $ref: '#/components/schemas/UserStory'
        comments:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
    UserStory:
      type: object
      required: [id, reference_id, title, status, priority, epic_id, creator_id, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        reference_id:
          type: string
          pattern: '^US-\d+$'
        title:
          type: string
          maxLength: 500
        description:
          type: string
          maxLength: 2000
        status:
          $ref: '#/components/schemas/UserStoryStatus'
        priority:
          $ref: '#/components/schemas/Priority'
        epic_id:
          type: string
          format: uuid
        creator_id:
          type: string
          format: uuid
        assignee_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        # Optional populated fields
        epic:
          $ref: '#/components/schemas/Epic'
        creator:
          $ref: '#/components/schemas/User'
        assignee:
          $ref: '#/components/schemas/User'
        acceptance_criteria:
          type: array
          items:
            $ref: '#/components/schemas/AcceptanceCriteria'
        requirements:
          type: array
          items:
            $ref: '#/components/schemas/Requirement'
        comments:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
...
    CreateUserStoryRequest:
      type: object
      required: [title, priority, epic_id]
      properties:
        title:
          type: string
          maxLength: 500
        description:
          type: string
          maxLength: 2000
        priority:
          $ref: '#/components/schemas/Priority'
        epic_id:
          type: string
          format: uuid
        assignee_id:
          type: string
          format: uuid
    UpdateUserStoryRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 500
        description:
          type: string
          maxLength: 2000
        priority:
          $ref: '#/components/schemas/Priority'
        assignee_id:
          type: string
          format: uuid
    CreateAcceptanceCriteriaRequest:
      type: object
      required: [description, user_story_id]
      properties:
        description:
          type: string
        user_story_id:
          type: string
          format: uuid
...
    UserStoryListResponse:
      allOf:
        - $ref: '#/components/schemas/ListResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/UserStory'
    AcceptanceCriteriaListResponse:
      allOf:
        - $ref: '#/components/schemas/ListResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/AcceptanceCriteria'
...